feat: move stronghold generation out of the agent into own runnable.
So that managing the stronghold can be done outside of the boot process
of the agent service.
So that we don't need to run the service twice with changing the config
inbetween in order to generate a new stronghold.
So that we can overwrite a stronghold if we want.
So that we can remove logic and config to generate the file from the
agent. It now only loads the file.

The runnable prints the stronghold config after generating a file. We
can lift the DID from that for use in the config when running the agent.

This makes for a more predictable runnable.

feat: Log passed in arguments when failing secret manager with arg error