feat: Replace impierce github actions with one action for our deployment
Impierce actions fail bc we lack access to the services used. And we can
rely on upstream running them for us.

We need one action to run: so that we can deploy on SDP.

So that in development we have a fully working configuration.
Only secrets must still be set in env.

Layer compose file so that we can rely on upstream and override in
our own